Remove gzip
As of this commit brotli compression is supported by 96.81% which is
more than some of the other features you tend to use with datastar
like popovers (89.51%).

Hyperlith's approach is generally around opinionated minimalism. By
only supporting one compression type we avoid writing code that
dispatched based on client compression support which has a cost in
terms of performance and maintenance.