Add etag support for shims
Using etag to cache shims, saves a few bytes and a bit of cpu when
users revisit the page and the shim hasn't changed.

Quirk with browsers is that cache settings are per URL not per
URL + METHOD this means that GET and POST cache headers can
mess with each other. To get around this an unused query param
is added to the url.

See https://httpwg.org/specs/rfc9111.html#rfc.section.2

... However, many HTTP caches in common use today only cache GET
responses and therefore only use the URI as the cache key ...